+++ Magna/src/main/java/de/morigm/magna/chat/Chat.java
@@ -43,4 +43,11 @@ public class Chat
 		com.sendMessage(prefix + no_online);
 	}
 	
+	public static String toFirstUpAndRemainderLower(String text)
+	{
+		String up = text.toUpperCase();
+		String lower = text.toLowerCase();
+		return up.substring(0, 1) + lower.substring(1);
+	}
+	
 }